Function introduced to compute the areas of the control volumes

+++ src/c++/2d/2d_wave/packages/include/bed_shear_stress.h
@@ -34,15 +34,15 @@
 #include <Eigen/IterativeLinearSolvers>
 #include <Eigen/Sparse>
 
-int bed_shear_stress_matrix_and_rhs(double* values, int row, int c_eq, int q_eq, int r_eq, Eigen::VectorXd& rhs,
+int bed_shear_stress_matrix_and_rhs(double* values, size_t row, size_t c_eq, size_t q_eq, size_t r_eq, Eigen::VectorXd& rhs,
     std::vector<double>& x, std::vector<double>& y,
     std::vector<double>& htheta, std::vector<double>& qtheta, std::vector<double>& rtheta,
-    double cf, double theta, int nx, int ny);
+    double cf, double theta, size_t nx, size_t ny);
 void bed_shear_stress_post_rhs(std::vector<double>& rhs_q, std::vector<double>& rhs_r, 
     std::vector<double>& hn, std::vector<double>& qn, std::vector<double>& rn,
-    double cf, int nx, int ny);
+    double cf, size_t nx, size_t ny);
 
-    inline int bed_shear_stress_idx(int i, int j, int ny);
-    inline void add_value(double * values, int col, double data);
+    inline size_t bed_shear_stress_idx(size_t i, size_t j, size_t ny);
+    inline void add_value(double * values, size_t col, double data);
 
 #endif  // __BEDSHEARSTRESS_H__
